Move close codes to separate module
